Refactor map_factory

+++ src/native/types.h
@@ -1,6 +1,6 @@
 #ifndef TYPE_H
 #define TYPE_H
 
-typedef unsigned long long index_t;
+typedef unsigned long index_t;
 
 #endif
